Bug fix
fixed an issue when base meshes could not be generated from cyclic
skeletons without branching aka toruses

+++ Lib/include/OpenMesh/Tools/Utils/HeapT.hh
@@ -1,7 +1,7 @@
 /*===========================================================================*\
  *                                                                           *
  *                               OpenMesh                                    *
- *      Copyright (C) 2001-2011 by Computer Graphics Group, RWTH Aachen      *
+ *      Copyright (C) 2001-2014 by Computer Graphics Group, RWTH Aachen      *
  *                           www.openmesh.org                                *
  *                                                                           *
  *---------------------------------------------------------------------------* 
@@ -34,8 +34,8 @@
 
 /*===========================================================================*\
  *                                                                           *             
- *   $Revision: 563 $                                                         *
- *   $Date: 2012-03-21 16:56:38 +0100 (Wed, 21 Mar 2012) $                   *
+ *   $Revision: 990 $                                                         *
+ *   $Date: 2014-02-05 10:01:07 +0100 (Mi, 05 Feb 2014) $                   *
  *                                                                           *
 \*===========================================================================*/
 
@@ -82,7 +82,7 @@ namespace Utils { // BEGIN_NS_UTILS
 
 
 /** This class demonstrates the HeapInterface's interface.  If you
- *  want to build your customized heap you will have to specity a heap
+ *  want to build your customized heap you will have to specify a heap
  *  interface class and use this class as a template parameter for the
  *  class HeapT. This class defines the interface that this heap
  *  interface has to implement.
@@ -111,8 +111,8 @@ struct HeapInterfaceT
  *
  *  An efficient, highly customizable heap.
  *
- *  The main difference (and performace boost) of this heap compared
- *  to e.g. the heap of the STL is that here to positions of the
+ *  The main difference (and performance boost) of this heap compared
+ *  to e.g. the heap of the STL is that here the positions of the
  *  heap's elements are accessible from the elements themself.
  *  Therefore if one changes the priority of an element one does not
  *  have to remove and re-insert this element, but can just call the
@@ -156,10 +156,10 @@ public:
   bool empty() const { return HeapVector::empty(); }
 
   /// returns the size of heap
-  unsigned int size() const { return HeapVector::size(); }
+  size_t size() const { return HeapVector::size(); }
 
   /// reserve space for _n entries
-  void reserve(unsigned int _n) { HeapVector::reserve(_n); }
+  void reserve(size_t _n) { HeapVector::reserve(_n); }
 
   /// reset heap position to -1 (not in heap)
   void reset_heap_position(HeapEntry _h)
@@ -244,13 +244,13 @@ public:
     {
       if (((j=left(i))<size()) && interface_.greater(entry(i), entry(j))) 
       {
-	omerr() << "Heap condition violated\n";
-	ok=false;
+        omerr() << "Heap condition violated\n";
+        ok=false;
       }
       if (((j=right(i))<size()) && interface_.greater(entry(i), entry(j)))
       {
-	omerr() << "Heap condition violated\n";
-	ok=false;
+        omerr() << "Heap condition violated\n";
+        ok=false;
       }
     }
     return ok;
@@ -266,15 +266,15 @@ private:
 
   
   /// Upheap. Establish heap property.
-  void upheap(unsigned int _idx);
+  void upheap(size_t _idx);
 
   
   /// Downheap. Establish heap property.
-  void downheap(unsigned int _idx);
+  void downheap(size_t _idx);
 
   
   /// Get the entry at index _idx
-  inline HeapEntry entry(unsigned int _idx) const
+  inline HeapEntry entry(size_t _idx) const
   {
     assert(_idx < size());
     return (Base::operator[](_idx));
@@ -282,20 +282,20 @@ private:
 
   
   /// Set entry _h to index _idx and update _h's heap position.
-  inline void entry(unsigned int _idx, HeapEntry _h) 
+  inline void entry(size_t _idx, HeapEntry _h)
   {
     assert(_idx < size());
     Base::operator[](_idx) = _h;
-    interface_.set_heap_position(_h, _idx);
+    interface_.set_heap_position(_h, int(_idx));
   }
 
   
   /// Get parent's index
-  inline unsigned int parent(unsigned int _i) { return (_i-1)>>1; }
+  inline size_t parent(size_t _i) { return (_i-1)>>1; }
   /// Get left child's index
-  inline unsigned int left(unsigned int _i)   { return (_i<<1)+1; }
+  inline size_t left(size_t _i)   { return (_i<<1)+1; }
   /// Get right child's index
-  inline unsigned int right(unsigned int _i)  { return (_i<<1)+2; }
+  inline size_t right(size_t _i)  { return (_i<<1)+2; }
 
 };
 
@@ -308,13 +308,12 @@ private:
 template <class HeapEntry, class HeapInterface>
 void
 HeapT<HeapEntry, HeapInterface>::
-upheap(unsigned int _idx)
+upheap(size_t _idx)
 {
   HeapEntry     h = entry(_idx);
-  unsigned int  parentIdx;
+  size_t        parentIdx;
 
-  while ((_idx>0) &&
-	 interface_.less(h, entry(parentIdx=parent(_idx))))
+  while ((_idx>0) && interface_.less(h, entry(parentIdx=parent(_idx))))
   {
     entry(_idx, entry(parentIdx));
     _idx = parentIdx;    
@@ -330,19 +329,18 @@ upheap(unsigned int _idx)
 template <class HeapEntry, class HeapInterface>
 void
 HeapT<HeapEntry, HeapInterface>::
-downheap(unsigned int _idx)
+downheap(size_t _idx)
 {
   HeapEntry     h = entry(_idx);
-  unsigned int  childIdx;
-  unsigned int  s = size();
+  size_t        childIdx;
+  size_t        s = size();
   
   while(_idx < s)
   {
     childIdx = left(_idx);
     if (childIdx >= s) break;
     
-    if ((childIdx+1 < s) &&
-	(interface_.less(entry(childIdx+1), entry(childIdx))))
+    if ((childIdx + 1 < s) && (interface_.less(entry(childIdx + 1), entry(childIdx))))
       ++childIdx;
     
     if (interface_.less(h, entry(childIdx))) break;
