+++ Lib/include/OpenMesh/Tools/Utils/StripifierT.hh
@@ -1,7 +1,7 @@
 /*===========================================================================*\
  *                                                                           *
  *                               OpenMesh                                    *
- *      Copyright (C) 2001-2011 by Computer Graphics Group, RWTH Aachen      *
+ *      Copyright (C) 2001-2014 by Computer Graphics Group, RWTH Aachen      *
  *                           www.openmesh.org                                *
  *                                                                           *
  *---------------------------------------------------------------------------* 
@@ -34,8 +34,8 @@
 
 /*===========================================================================*\
  *                                                                           *             
- *   $Revision: 362 $                                                         *
- *   $Date: 2011-01-26 10:21:12 +0100 (Wed, 26 Jan 2011) $                   *
+ *   $Revision: 990 $                                                         *
+ *   $Date: 2014-02-05 10:01:07 +0100 (Mi, 05 Feb 2014) $                   *
  *                                                                           *
 \*===========================================================================*/
 
@@ -92,13 +92,13 @@ public:
   ~StripifierT();
 
   /// Compute triangle strips, returns number of strips
-  unsigned int stripify();
+  size_t stripify();
 
   /// delete all strips
   void clear() { Strips().swap(strips_); }
 
   /// returns number of strips
-  unsigned int n_strips() const { return strips_.size(); }
+  size_t n_strips() const { return strips_.size(); }
 
   /// are strips computed?
   bool is_valid() const { return !strips_.empty(); }
